900 Changes!

+++ src/node/linearextrudenode.cpp
@@ -1,6 +1,6 @@
 /*
  *   RapCAD - Rapid prototyping CAD IDE (www.rapcad.org)
- *   Copyright (C) 2010-2011 Giles Bathgate
+ *   Copyright (C) 2010-2014 Giles Bathgate
  *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
@@ -22,12 +22,12 @@ LinearExtrudeNode::LinearExtrudeNode()
 {
 }
 
-void LinearExtrudeNode::setHeight(double h)
+void LinearExtrudeNode::setHeight(decimal h)
 {
 	height=h;
 }
 
-double LinearExtrudeNode::getHeight() const
+decimal LinearExtrudeNode::getHeight() const
 {
 	return height;
 }
