+++ src/rangeiterator.cpp
@@ -1,6 +1,6 @@
 /*
  *   RapCAD - Rapid prototyping CAD IDE (www.rapcad.org)
- *   Copyright (C) 2010-2011 Giles Bathgate
+ *   Copyright (C) 2010-2014 Giles Bathgate
  *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
@@ -30,7 +30,7 @@ RangeIterator::RangeIterator(RangeValue* rng)
 
 	step=range->getStep();
 	if(!step) {
-		double i=reverse?-1.0:1.0;
+		decimal i=reverse?-1.0:1.0;
 		defaultStep=new NumberValue(i);
 		step=defaultStep;
 	} else {
